+++ app/src/main/java/com/motologr/ui/data/objects/maint/Repair.kt
@@ -1,9 +1,10 @@
 package com.motologr.ui.data.objects.maint
 
 import com.motologr.ui.data.logging.Loggable
+import java.math.BigDecimal
 import java.util.Date
 
-class Repair(var repairType: Int, var price: Double, var repairDate: Date, var repairProvider: String, var comment: String) : Loggable(repairDate, 0, price) {
+class Repair(var repairType: Int, var price: BigDecimal, var repairDate: Date, var repairProvider: String, var comment: String) : Loggable(repairDate, 0, price) {
     fun returnRepairType() : String {
         return when (repairType) {
             0 -> {
