+++ app/src/main/java/com/motologr/ui/data/objects/maint/Wof.kt
@@ -1,6 +1,7 @@
 package com.motologr.ui.data.objects.maint
 
 import com.motologr.ui.data.logging.Loggable
+import java.math.BigDecimal
 import java.util.Date
 
-class Wof(var wofDate: Date, var wofCompletedDate: Date, var price: Double) : Loggable(wofCompletedDate, 2, price)
+class Wof(var wofDate: Date, var wofCompletedDate: Date, var price: BigDecimal) : Loggable(wofCompletedDate, 2, price)
