+++ app/src/main/java/com/motologr/ui/data/objects/reg/Reg.kt
@@ -1,6 +1,7 @@
 package com.motologr.ui.data.objects.reg
 
 import com.motologr.ui.data.logging.Loggable
+import java.math.BigDecimal
 import java.util.Date
 
-class Reg(var newRegExpiryDate: Date, var regExpiryDate: Date, var monthsExtended: Int, var price: Double) : Loggable(regExpiryDate, 3, price)
+class Reg(var newRegExpiryDate: Date, var regExpiryDate: Date, var monthsExtended: Int, var price: BigDecimal) : Loggable(regExpiryDate, 3, price)
